# api.yml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
tags:
  - name: basic-error-controller
    description: Basic Error Controller
  - name: meetup-controller
    description: Meetup Controller
  - name: users-controller
    description: Users Controller
paths:
  /error:
    get:
      tags:
        - basic-error-controller
      summary: errorHtml
      operationId: errorHtmlUsingGET
      produces:
        - text/html
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    head:
      tags:
        - basic-error-controller
      summary: errorHtml
      operationId: errorHtmlUsingHEAD
      consumes:
        - application/json
      produces:
        - text/html
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags:
        - basic-error-controller
      summary: errorHtml
      operationId: errorHtmlUsingPOST
      consumes:
        - application/json
      produces:
        - text/html
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - basic-error-controller
      summary: errorHtml
      operationId: errorHtmlUsingPUT
      consumes:
        - application/json
      produces:
        - text/html
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - basic-error-controller
      summary: errorHtml
      operationId: errorHtmlUsingDELETE
      produces:
        - text/html
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    options:
      tags:
        - basic-error-controller
      summary: errorHtml
      operationId: errorHtmlUsingOPTIONS
      consumes:
        - application/json
      produces:
        - text/html
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    patch:
      tags:
        - basic-error-controller
      summary: errorHtml
      operationId: errorHtmlUsingPATCH
      consumes:
        - application/json
      produces:
        - text/html
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ModelAndView'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  '/meetup/add-user/{meetup-id}':
    post:
      tags:
        - meetup-controller
      summary: addUser
      operationId: addUserUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: meetup-id
          in: path
          description: meetup-id
          required: true
          type: string
        - name: user-id
          in: query
          description: user-id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
            enum:
              - ACCEPTED
              - ALREADY_REPORTED
              - BAD_GATEWAY
              - BAD_REQUEST
              - BANDWIDTH_LIMIT_EXCEEDED
              - CHECKPOINT
              - CONFLICT
              - CONTINUE
              - CREATED
              - DESTINATION_LOCKED
              - EXPECTATION_FAILED
              - FAILED_DEPENDENCY
              - FORBIDDEN
              - FOUND
              - GATEWAY_TIMEOUT
              - GONE
              - HTTP_VERSION_NOT_SUPPORTED
              - IM_USED
              - INSUFFICIENT_SPACE_ON_RESOURCE
              - INSUFFICIENT_STORAGE
              - INTERNAL_SERVER_ERROR
              - I_AM_A_TEAPOT
              - LENGTH_REQUIRED
              - LOCKED
              - LOOP_DETECTED
              - METHOD_FAILURE
              - METHOD_NOT_ALLOWED
              - MOVED_PERMANENTLY
              - MOVED_TEMPORARILY
              - MULTIPLE_CHOICES
              - MULTI_STATUS
              - NETWORK_AUTHENTICATION_REQUIRED
              - NON_AUTHORITATIVE_INFORMATION
              - NOT_ACCEPTABLE
              - NOT_EXTENDED
              - NOT_FOUND
              - NOT_IMPLEMENTED
              - NOT_MODIFIED
              - NO_CONTENT
              - OK
              - PARTIAL_CONTENT
              - PAYLOAD_TOO_LARGE
              - PAYMENT_REQUIRED
              - PERMANENT_REDIRECT
              - PRECONDITION_FAILED
              - PRECONDITION_REQUIRED
              - PROCESSING
              - PROXY_AUTHENTICATION_REQUIRED
              - REQUESTED_RANGE_NOT_SATISFIABLE
              - REQUEST_ENTITY_TOO_LARGE
              - REQUEST_HEADER_FIELDS_TOO_LARGE
              - REQUEST_TIMEOUT
              - REQUEST_URI_TOO_LONG
              - RESET_CONTENT
              - SEE_OTHER
              - SERVICE_UNAVAILABLE
              - SWITCHING_PROTOCOLS
              - TEMPORARY_REDIRECT
              - TOO_EARLY
              - TOO_MANY_REQUESTS
              - UNAUTHORIZED
              - UNAVAILABLE_FOR_LEGAL_REASONS
              - UNPROCESSABLE_ENTITY
              - UNSUPPORTED_MEDIA_TYPE
              - UPGRADE_REQUIRED
              - URI_TOO_LONG
              - USE_PROXY
              - VARIANT_ALSO_NEGOTIATES
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /meetup/all:
    get:
      tags:
        - meetup-controller
      summary: getMeetups
      operationId: getMeetupsUsingGET
      produces:
        - '*/*'
      parameters:
        - name: user_id
          in: query
          description: user_id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MeetupDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /meetup/create:
    post:
      tags:
        - meetup-controller
      summary: create
      operationId: createUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: meetupDto
          description: meetupDto
          required: true
          schema:
            $ref: '#/definitions/MeetupDto'
      responses:
        '200':
          description: OK
          schema:
            type: string
            enum:
              - ACCEPTED
              - ALREADY_REPORTED
              - BAD_GATEWAY
              - BAD_REQUEST
              - BANDWIDTH_LIMIT_EXCEEDED
              - CHECKPOINT
              - CONFLICT
              - CONTINUE
              - CREATED
              - DESTINATION_LOCKED
              - EXPECTATION_FAILED
              - FAILED_DEPENDENCY
              - FORBIDDEN
              - FOUND
              - GATEWAY_TIMEOUT
              - GONE
              - HTTP_VERSION_NOT_SUPPORTED
              - IM_USED
              - INSUFFICIENT_SPACE_ON_RESOURCE
              - INSUFFICIENT_STORAGE
              - INTERNAL_SERVER_ERROR
              - I_AM_A_TEAPOT
              - LENGTH_REQUIRED
              - LOCKED
              - LOOP_DETECTED
              - METHOD_FAILURE
              - METHOD_NOT_ALLOWED
              - MOVED_PERMANENTLY
              - MOVED_TEMPORARILY
              - MULTIPLE_CHOICES
              - MULTI_STATUS
              - NETWORK_AUTHENTICATION_REQUIRED
              - NON_AUTHORITATIVE_INFORMATION
              - NOT_ACCEPTABLE
              - NOT_EXTENDED
              - NOT_FOUND
              - NOT_IMPLEMENTED
              - NOT_MODIFIED
              - NO_CONTENT
              - OK
              - PARTIAL_CONTENT
              - PAYLOAD_TOO_LARGE
              - PAYMENT_REQUIRED
              - PERMANENT_REDIRECT
              - PRECONDITION_FAILED
              - PRECONDITION_REQUIRED
              - PROCESSING
              - PROXY_AUTHENTICATION_REQUIRED
              - REQUESTED_RANGE_NOT_SATISFIABLE
              - REQUEST_ENTITY_TOO_LARGE
              - REQUEST_HEADER_FIELDS_TOO_LARGE
              - REQUEST_TIMEOUT
              - REQUEST_URI_TOO_LONG
              - RESET_CONTENT
              - SEE_OTHER
              - SERVICE_UNAVAILABLE
              - SWITCHING_PROTOCOLS
              - TEMPORARY_REDIRECT
              - TOO_EARLY
              - TOO_MANY_REQUESTS
              - UNAUTHORIZED
              - UNAVAILABLE_FOR_LEGAL_REASONS
              - UNPROCESSABLE_ENTITY
              - UNSUPPORTED_MEDIA_TYPE
              - UPGRADE_REQUIRED
              - URI_TOO_LONG
              - USE_PROXY
              - VARIANT_ALSO_NEGOTIATES
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /meetup/temperature:
    get:
      tags:
        - meetup-controller
      summary: getTemperature
      operationId: getTemperatureUsingGET
      produces:
        - '*/*'
      parameters:
        - name: meetup_id
          in: query
          description: meetup_id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: integer
            format: int32
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /meetup/temperature-and-beers:
    get:
      tags:
        - meetup-controller
      summary: getTempAndBeers
      operationId: getTempAndBeersUsingGET
      produces:
        - '*/*'
      parameters:
        - name: meetup_id
          in: query
          description: meetup_id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MeetupInformationDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/meetup/{id}':
    get:
      tags:
        - meetup-controller
      summary: getMeetup
      operationId: getMeetupUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MeetupDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /users:
    post:
      tags:
        - users-controller
      summary: create
      operationId: createUsingPOST_1
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: userDto
          description: userDto
          required: true
          schema:
            $ref: '#/definitions/UserDto'
      responses:
        '200':
          description: OK
          schema:
            type: string
            enum:
              - ACCEPTED
              - ALREADY_REPORTED
              - BAD_GATEWAY
              - BAD_REQUEST
              - BANDWIDTH_LIMIT_EXCEEDED
              - CHECKPOINT
              - CONFLICT
              - CONTINUE
              - CREATED
              - DESTINATION_LOCKED
              - EXPECTATION_FAILED
              - FAILED_DEPENDENCY
              - FORBIDDEN
              - FOUND
              - GATEWAY_TIMEOUT
              - GONE
              - HTTP_VERSION_NOT_SUPPORTED
              - IM_USED
              - INSUFFICIENT_SPACE_ON_RESOURCE
              - INSUFFICIENT_STORAGE
              - INTERNAL_SERVER_ERROR
              - I_AM_A_TEAPOT
              - LENGTH_REQUIRED
              - LOCKED
              - LOOP_DETECTED
              - METHOD_FAILURE
              - METHOD_NOT_ALLOWED
              - MOVED_PERMANENTLY
              - MOVED_TEMPORARILY
              - MULTIPLE_CHOICES
              - MULTI_STATUS
              - NETWORK_AUTHENTICATION_REQUIRED
              - NON_AUTHORITATIVE_INFORMATION
              - NOT_ACCEPTABLE
              - NOT_EXTENDED
              - NOT_FOUND
              - NOT_IMPLEMENTED
              - NOT_MODIFIED
              - NO_CONTENT
              - OK
              - PARTIAL_CONTENT
              - PAYLOAD_TOO_LARGE
              - PAYMENT_REQUIRED
              - PERMANENT_REDIRECT
              - PRECONDITION_FAILED
              - PRECONDITION_REQUIRED
              - PROCESSING
              - PROXY_AUTHENTICATION_REQUIRED
              - REQUESTED_RANGE_NOT_SATISFIABLE
              - REQUEST_ENTITY_TOO_LARGE
              - REQUEST_HEADER_FIELDS_TOO_LARGE
              - REQUEST_TIMEOUT
              - REQUEST_URI_TOO_LONG
              - RESET_CONTENT
              - SEE_OTHER
              - SERVICE_UNAVAILABLE
              - SWITCHING_PROTOCOLS
              - TEMPORARY_REDIRECT
              - TOO_EARLY
              - TOO_MANY_REQUESTS
              - UNAUTHORIZED
              - UNAVAILABLE_FOR_LEGAL_REASONS
              - UNPROCESSABLE_ENTITY
              - UNSUPPORTED_MEDIA_TYPE
              - UPGRADE_REQUIRED
              - URI_TOO_LONG
              - USE_PROXY
              - VARIANT_ALSO_NEGOTIATES
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /users/auth:
    post:
      tags:
        - users-controller
      summary: authentication
      operationId: authenticationUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: userLogin
          description: userLogin
          required: true
          schema:
            $ref: '#/definitions/UserLogin'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/users/checkin-meetup/{user-id}':
    post:
      tags:
        - users-controller
      summary: inscribeToMeetup
      operationId: inscribeToMeetupUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: meetup_id
          in: query
          description: meetup_id
          required: true
          type: string
        - name: user-id
          in: path
          description: user-id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
            enum:
              - ACCEPTED
              - ALREADY_REPORTED
              - BAD_GATEWAY
              - BAD_REQUEST
              - BANDWIDTH_LIMIT_EXCEEDED
              - CHECKPOINT
              - CONFLICT
              - CONTINUE
              - CREATED
              - DESTINATION_LOCKED
              - EXPECTATION_FAILED
              - FAILED_DEPENDENCY
              - FORBIDDEN
              - FOUND
              - GATEWAY_TIMEOUT
              - GONE
              - HTTP_VERSION_NOT_SUPPORTED
              - IM_USED
              - INSUFFICIENT_SPACE_ON_RESOURCE
              - INSUFFICIENT_STORAGE
              - INTERNAL_SERVER_ERROR
              - I_AM_A_TEAPOT
              - LENGTH_REQUIRED
              - LOCKED
              - LOOP_DETECTED
              - METHOD_FAILURE
              - METHOD_NOT_ALLOWED
              - MOVED_PERMANENTLY
              - MOVED_TEMPORARILY
              - MULTIPLE_CHOICES
              - MULTI_STATUS
              - NETWORK_AUTHENTICATION_REQUIRED
              - NON_AUTHORITATIVE_INFORMATION
              - NOT_ACCEPTABLE
              - NOT_EXTENDED
              - NOT_FOUND
              - NOT_IMPLEMENTED
              - NOT_MODIFIED
              - NO_CONTENT
              - OK
              - PARTIAL_CONTENT
              - PAYLOAD_TOO_LARGE
              - PAYMENT_REQUIRED
              - PERMANENT_REDIRECT
              - PRECONDITION_FAILED
              - PRECONDITION_REQUIRED
              - PROCESSING
              - PROXY_AUTHENTICATION_REQUIRED
              - REQUESTED_RANGE_NOT_SATISFIABLE
              - REQUEST_ENTITY_TOO_LARGE
              - REQUEST_HEADER_FIELDS_TOO_LARGE
              - REQUEST_TIMEOUT
              - REQUEST_URI_TOO_LONG
              - RESET_CONTENT
              - SEE_OTHER
              - SERVICE_UNAVAILABLE
              - SWITCHING_PROTOCOLS
              - TEMPORARY_REDIRECT
              - TOO_EARLY
              - TOO_MANY_REQUESTS
              - UNAUTHORIZED
              - UNAVAILABLE_FOR_LEGAL_REASONS
              - UNPROCESSABLE_ENTITY
              - UNSUPPORTED_MEDIA_TYPE
              - UPGRADE_REQUIRED
              - URI_TOO_LONG
              - USE_PROXY
              - VARIANT_ALSO_NEGOTIATES
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/users/{id}':
    get:
      tags:
        - users-controller
      summary: getUser
      operationId: getUserUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/users/{id}/meetups-checkin':
    get:
      tags:
        - users-controller
      summary: getCheckinMeetups
      operationId: getCheckinMeetupsUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MeetupDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/users/{id}/meetups-created':
    get:
      tags:
        - users-controller
      summary: getMeetupsCreated
      operationId: getMeetupsCreatedUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MeetupDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/users/{id}/meetups-inscribed':
    get:
      tags:
        - users-controller
      summary: getMeetupsInscribed
      operationId: getMeetupsInscribedUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MeetupDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  MeetupDto:
    type: object
    properties:
      beers:
        type: integer
        format: int32
      city:
        type: string
      creatorId:
        type: integer
        format: int64
      date:
        type: string
        format: date-time
      duration:
        type: integer
        format: int32
      enrolled:
        type: integer
        format: int32
      id:
        type: integer
        format: int64
      name:
        type: string
    title: MeetupDto
  MeetupInformationDto:
    type: object
    properties:
      beers:
        type: integer
        format: int32
      description:
        type: string
      temp:
        type: integer
        format: int32
    title: MeetupInformationDto
  ModelAndView:
    type: object
    properties:
      empty:
        type: boolean
      model:
        type: object
      modelMap:
        type: object
        additionalProperties:
          type: object
      reference:
        type: boolean
      status:
        type: string
        enum:
          - ACCEPTED
          - ALREADY_REPORTED
          - BAD_GATEWAY
          - BAD_REQUEST
          - BANDWIDTH_LIMIT_EXCEEDED
          - CHECKPOINT
          - CONFLICT
          - CONTINUE
          - CREATED
          - DESTINATION_LOCKED
          - EXPECTATION_FAILED
          - FAILED_DEPENDENCY
          - FORBIDDEN
          - FOUND
          - GATEWAY_TIMEOUT
          - GONE
          - HTTP_VERSION_NOT_SUPPORTED
          - IM_USED
          - INSUFFICIENT_SPACE_ON_RESOURCE
          - INSUFFICIENT_STORAGE
          - INTERNAL_SERVER_ERROR
          - I_AM_A_TEAPOT
          - LENGTH_REQUIRED
          - LOCKED
          - LOOP_DETECTED
          - METHOD_FAILURE
          - METHOD_NOT_ALLOWED
          - MOVED_PERMANENTLY
          - MOVED_TEMPORARILY
          - MULTIPLE_CHOICES
          - MULTI_STATUS
          - NETWORK_AUTHENTICATION_REQUIRED
          - NON_AUTHORITATIVE_INFORMATION
          - NOT_ACCEPTABLE
          - NOT_EXTENDED
          - NOT_FOUND
          - NOT_IMPLEMENTED
          - NOT_MODIFIED
          - NO_CONTENT
          - OK
          - PARTIAL_CONTENT
          - PAYLOAD_TOO_LARGE
          - PAYMENT_REQUIRED
          - PERMANENT_REDIRECT
          - PRECONDITION_FAILED
          - PRECONDITION_REQUIRED
          - PROCESSING
          - PROXY_AUTHENTICATION_REQUIRED
          - REQUESTED_RANGE_NOT_SATISFIABLE
          - REQUEST_ENTITY_TOO_LARGE
          - REQUEST_HEADER_FIELDS_TOO_LARGE
          - REQUEST_TIMEOUT
          - REQUEST_URI_TOO_LONG
          - RESET_CONTENT
          - SEE_OTHER
          - SERVICE_UNAVAILABLE
          - SWITCHING_PROTOCOLS
          - TEMPORARY_REDIRECT
          - TOO_EARLY
          - TOO_MANY_REQUESTS
          - UNAUTHORIZED
          - UNAVAILABLE_FOR_LEGAL_REASONS
          - UNPROCESSABLE_ENTITY
          - UNSUPPORTED_MEDIA_TYPE
          - UPGRADE_REQUIRED
          - URI_TOO_LONG
          - USE_PROXY
          - VARIANT_ALSO_NEGOTIATES
      view:
        $ref: '#/definitions/View'
      viewName:
        type: string
    title: ModelAndView
  UserDto:
    type: object
    properties:
      age:
        type: integer
        format: int32
      first_name:
        type: string
      id:
        type: integer
        format: int64
      last_name:
        type: string
      password:
        type: string
      role:
        type: string
        enum:
          - ADMIN
          - USER
      sur_name:
        type: string
    title: UserDto
  UserLogin:
    type: object
    properties:
      password:
        type: string
      username:
        type: string
    title: UserLogin
  View:
    type: object
    properties:
      contentType:
        type: string
    title: View