# src/main/java/com/api/beer/beers/users/dto/UserLogin.java
package com.api.beer.beers.users.dto;

public class UserLogin {

    private String username;
    private String password;

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}